requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
decorator==4.4.2
idna==3.10
mutagen==1.47.0
python-dotenv==1.1.1
requests==2.32.4
spotipy==2.25.1
tqdm==4.67.1
urllib3==2.5.0
yt-dlp==2025.6.30
